Rename event components from meetup to events & organize
