+++ client/components/events/eventsListItem.js
